// packages/tauri-api-adapter-v1/src/api/client-types.ts
export interface INetwork {
  getInterfaces: typeof getInterfaces
  getNonEmptyInterfaces: typeof getNonEmptyInterfaces
  findAvailablePort: typeof findAvailablePort
  isPortTaken: typeof isPortTaken
  isHttpPortOpen: typeof isHttpPortOpen
  // scanOnlineIpPortPairs: typeof scanOnlineIpPortPairs
  // scanOnlineIpsByPort: typeof scanOnlineIpsByPort
  // nonLocalhostNetworks: typeof nonLocalhostNetworks
  // localServerIsRunning: typeof localServerIsRunning
  // scanLocalNetworkOnlineHostsByPort: typeof scanLocalNetworkOnlineHostsByPort
}